refactor: organize layers folder
